requirements.txt
aiohttp~=3.12.14
python-dotenv~=1.1.1
discord.py~=2.5.2
flask~=3.1.1
requests~=2.32.4
SQLAlchemy~=2.0.41